pyproject.toml
[tool.poetry]
name = "dspace-reports"
version = "2.0.0"
description = "A python3-based tool to generate and email views and downloads statistical reports for a DSpace 7+ repository."
authors = ["Nicholas Woodward <njw@austin.utexas.edu>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
certifi = "2024.7.4"
chardet = "5.2.0"
idna = "3.7"
isort = "5.13.2"
lazy-object-proxy = "1.10.0"
mccabe = "==0.7.0"
psycopg = "3.2.1"
pylint = "3.2.5"
python-dateutil = "2.9.0.post0"
PyYAML = "6.0.2"
requests = "2.32.3"
six = "==1.16.0"
toml = "==0.10.2"
urllib3 = "2.2.2"
wrapt = "1.16.0"
XlsxWriter = "3.2.0"
astroid = "3.2.3"
tomlkit = "0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
